Code cleanup and disabling assembly manager
