+++ src/Editor.meta
@@ -1,8 +1,3 @@
-fileFormatVersion: 2
-guid: 3f77bd7bac843b45d851f83be3888f20
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 4ce1edd77e654b50aed6e9fd803fe4f1
+timeCreated: 1633838389
