+++ src/Editor/Appalachia.CI.Packaging.Editor.asmdef
@@ -1,8 +1,9 @@
 {
-    "name": "Appalachia.CI.Packaging",
+    "name": "Appalachia.CI.Packaging.Editor",
     "rootNamespace": "",
     "references": [
-        "Appalachia.CI.SemVer"
+        "GUID:56ca8901b63fd204692a4a27f67b605e",
+        "GUID:c7f2f2911de381941be7186214305448"
     ],
     "includePlatforms": [
         "Editor"
