+++ src/Editor/PackageRegistry/Core/NPMCredential.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 854a13e69de04a8fa58a5df2f3deaea4
+timeCreated: 1633835270
