+++ src/Editor/PackageRegistry/NPM/ExpectContinueAware.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e7554137f0224ca9a8499c447d58997b
+timeCreated: 1633835270
