+++ src/Editor/PackageRegistry/NPM/NPMLoginRequest.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Appalachia.CI.Packaging.Editor.PackageRegistry.NPM
+{
+    [Serializable]
+    internal class NPMLoginRequest
+    {
+        public string name;
+        public string password;
+    }
+}
