+++ src/Editor/PackageRegistry/NPM/NPMLoginRequest.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6c7f2891a9b94238b758b0399d5107ca
+timeCreated: 1633835270
