+++ src/Editor/PackageRegistry/UI/CredentialManagerSettingsProvider.cs
@@ -1,13 +1,13 @@
 using System;
 using System.Collections.Generic;
-using Appalachia.CI.Packaging.PackageRegistry.Core;
+using Appalachia.CI.Packaging.Editor.PackageRegistry.Core;
 using UnityEditor;
 using UnityEditor.PackageManager;
 using UnityEditor.SceneManagement;
 using UnityEditorInternal;
 using UnityEngine;
 
-namespace Appalachia.CI.Packaging.PackageRegistry.UI
+namespace Appalachia.CI.Packaging.Editor.PackageRegistry.UI
 {
     internal static class CredentialManagerSettingsProvider
     {
