+++ src/Editor/PackageRegistry/UI/CredentialManagerSettingsProvider.cs.meta
