+++ src/Editor/PackageRegistry/UI/GetTokenView.cs
@@ -1,30 +1,10 @@
-using Appalachia.CI.Packaging.PackageRegistry.Core;
-using Appalachia.CI.Packaging.PackageRegistry.NPM;
+using Appalachia.CI.Packaging.Editor.PackageRegistry.Core;
+using Appalachia.CI.Packaging.Editor.PackageRegistry.NPM;
 using UnityEditor;
 using UnityEngine;
 
-namespace Appalachia.CI.Packaging.PackageRegistry.UI
+namespace Appalachia.CI.Packaging.Editor.PackageRegistry.UI
 {
-    internal class TokenMethod : GUIContent
-    {
-        internal GetToken action;
-        internal string passwordName;
-        internal string usernameName;
-
-        public TokenMethod(
-            string name,
-            string usernameName,
-            string passwordName,
-            GetToken action) : base(name)
-        {
-            this.usernameName = usernameName;
-            this.passwordName = passwordName;
-            this.action = action;
-        }
-
-        internal delegate bool GetToken(ScopedRegistry registry, string username, string password);
-    }
-
     internal class GetTokenView : EditorWindow
     {
         private static readonly TokenMethod[] methods =
