+++ src/Editor/PackageRegistry/UI/ScopedRegistryEditorView.cs
@@ -1,9 +1,9 @@
-using Appalachia.CI.Packaging.PackageRegistry.Core;
+using Appalachia.CI.Packaging.Editor.PackageRegistry.Core;
 using UnityEditor;
 using UnityEditorInternal;
 using UnityEngine;
 
-namespace Appalachia.CI.Packaging.PackageRegistry.UI
+namespace Appalachia.CI.Packaging.Editor.PackageRegistry.UI
 {
     internal class ScopedRegistryEditorView : EditorWindow
     {
