--- src/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "com.halodi.halodi-unity-package-registry-manager",
-  "version": "0.1.5",
-  "displayName": "Halodi Package Registry Manager",
-  "type": "tool",
-  "samples": [],
-  "description": "Simple tool to manage scoped registries and NPM credentials.",
-  "license": "Apache-2.0",
-  "author": {
-    "name": "Halodi",
-    "email": "jesper@halodi.com",
-    "url": "https://halodi.com"
-  },
-  "dependencies": {
-    "com.unity.nuget.newtonsoft-json": "2.0.0",
-    "com.unity.sharp-zip-lib": "1.2.2-preview.1"
-  }
-}
