--- src/package.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4480134b80472dc1a847e756ab3917cc
-PackageManifestImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
